meson: remove src/ from the libraries include directories
The '.' (src/) include directory was added by Peter Eszlari in his
inital port of liburing to meson in 253d497.

Exporting headers inside src/ which are clearly not meant to be exported
may shadow relevant system headers like 'syscall.h' breaking builds using
liburing as a meson subproject.

It is not included in the wrapdb downstream versions of liburing.
Fixes: 253d497

Signed-off-by: Florian Fischer <florian.fischer@muhq.space>